# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SCO Lead Management API
  version: 1.0.0
  description: >
    API untuk mengelola data EO/Mitra (Lead), assignment PIC, tracking status pipeline,
    dan dokumentasi aktivitas follow-up.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Role:
      type: string
      enum: [SUPERADMIN, SUPERIOR, PIC]
    PipelineStatus:
      type: string
      enum: [NEW_LEAD, CONTACTED, IN_DISCUSSION, PITCHING, NEGOTIATION, ON_HOLD, DEAL, LOST]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nama:
          type: string
        employeeId:
          type: string
        phoneNumber:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        isDeleted:
          type: boolean
        deletedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Lead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        namaEo:
          type: string
        tipeId:
          type: string
          format: uuid
        alamat:
          type: string
        speciality:
          type: string
          nullable: true
        linkSosmed:
          type: string
          nullable: true
        lastActivityDate:
          type: string
          format: date-time
          nullable: true
        lastActivityType:
          type: string
          nullable: true
        isDeleted:
          type: boolean
        deletedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
...
    LeadStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        leadId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PipelineStatus'
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    CreateUserRequest:
      type: object
      required: [nama, employee_id, phone_number, role]
      properties:
        nama:
          type: string
          minLength: 1
          maxLength: 255
        employee_id:
          type: string
          minLength: 1
          maxLength: 50
        phone_number:
          type: string
          minLength: 1
          maxLength: 20
        role:
          $ref: '#/components/schemas/Role'
    UpdateUserRequest:
      type: object
      properties:
        nama:
          type: string
          maxLength: 255
        employee_id:
          type: string
          maxLength: 50
        phone_number:
          type: string
          maxLength: 20
        role:
          $ref: '#/components/schemas/Role'
    CreateLeadRequest:
      type: object
      required: [nama_eo, tipe_id, alamat]
      properties:
        nama_eo:
          type: string
          minLength: 1
          maxLength: 255
        tipe_id:
          type: string
          format: uuid
        alamat:
          type: string
          minLength: 1
          maxLength: 500
        speciality:
          type: string
          maxLength: 255
        link_sosmed:
          type: string
          format: uri
    UpdateLeadRequest:
      type: object
      properties:
        nama_eo:
          type: string
          maxLength: 255
        tipe_id:
          type: string
          format: uuid
        alamat:
          type: string
          maxLength: 500
        speciality:
          type: string
          maxLength: 255
        link_sosmed:
          type: string
          format: uri
...
    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/PipelineStatus'
...
    PaginatedLeadResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Lead'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    Unauthorized:
      description: Token tidak valid atau tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Token tidak valid atau tidak ditemukan
    Forbidden:
      description: Akses ditolak
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Anda tidak memiliki akses untuk operasi ini
    NotFound:
      description: Resource tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource tidak ditemukan
    ValidationError:
      description: Data tidak valid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Data tidak valid
              details: {}
    Conflict:
      description: Data duplikat
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: CONFLICT
              message: Data sudah ada
paths:
...
  /api/users:
    post:
      tags: [Users]
      summary: Buat user baru (Superadmin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User berhasil dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [Users]
      summary: Daftar semua user (Superadmin only)
      responses:
        '200':
          description: Daftar user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Users]
      summary: Detail user (Superadmin only)
      responses:
        '200':
          description: Detail user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Update user (Superadmin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User berhasil diupdate
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: Soft delete user (Superadmin only)
      responses:
        '200':
          description: User berhasil dihapus
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: User berhasil dihapus
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Lead Types ====================
...
  /api/leads:
    post:
      tags: [Leads]
      summary: Buat Lead baru (Superadmin, Superior, PIC)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeadRequest'
      responses:
        '201':
          description: Lead berhasil dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Lead'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [Leads]
      summary: Daftar Lead dengan filter dan search
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PipelineStatus'
        - name: pic_id
          in: query
          schema:
            type: string
            format: uuid
        - name: tipe_id
          in: query
          schema:
            type: string
            format: uuid
        - name: last_activity_from
          in: query
          schema:
            type: string
            format: date-time
        - name: last_activity_to
          in: query
          schema:
            type: string
            format: date-time
        - name: search
          in: query
          description: Search nama_eo (case-insensitive)
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Daftar Lead (paginated)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLeadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/leads/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Leads]
      summary: Detail Lead
      responses:
        '200':
          description: Detail Lead
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Lead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Leads]
      summary: Update Lead (Superadmin, Superior)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeadRequest'
      responses:
        '200':
          description: Lead berhasil diupdate
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Lead'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Leads]
      summary: Soft delete Lead (Superadmin only)
      responses:
        '200':
          description: Lead berhasil dihapus
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Lead berhasil dihapus
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Contacts ====================
...
  /api/leads/{leadId}/status:
    parameters:
      - name: leadId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Status]
      summary: Update status Lead
      description: >
        Transisi status harus mengikuti state machine yang valid.
        Lead harus memiliki minimal 1 Activity sebelum status bisa diupdate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '201':
          description: Status berhasil diupdate
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/LeadStatus'
        '400':
          description: Transisi tidak valid atau belum ada Activity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                invalidTransition:
                  summary: Transisi status tidak valid
                  value:
                    success: false
                    error:
                      code: INVALID_TRANSITION
                      message: Transisi status tidak valid
                activityRequired:
                  summary: Belum ada Activity
                  value:
                    success: false
                    error:
                      code: ACTIVITY_REQUIRED
                      message: Lead harus memiliki minimal 1 Activity
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Status]
      summary: Histori status Lead
      responses:
        '200':
          description: Histori status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeadStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== Activities ====================
...